Improved stability when calling the "AsyncMqttClient::connect()" method multiple times.
Problem: When repeatedly calling this method, while the connection has not yet been established,
at the level of the "ESP Async TCP" library or at the level of Espressif SDK,
many "half-dead" objects are created that are referenced by the same callback functions.
These objects are not participating in the data exchange, however,
they periodically call these callback functions, which leads to malfunctions.

Solution: Adding a blocking flag "AsyncMqttClient::_lockMultyConnections" to the "Async MQTT Client" library,
which prevents consequences of multiple calls to the methods "AsyncMqttClient::connect()" and "AsyncMqttClient::disconnect()"