Initial version of websocket client filter for MQTT protocol
This is a generic websocket filter class that has no knowledge of where
the stream data is coming from or how to transmit into a stream. Actual
I/O is delegated to the caller. The constructor receives optional buffer
sizes, which sets the maximum frame length that can be generated by the
implementation. However, there is no maximum frame length limit - the
class allows streaming of the RX frame data in chunks. Websocket
protocol negotiation is supported but chosen protocol is not (yet)
enforced or reported.